+++ config/test/clippy.js
@@ -1,13 +1,11 @@
 export default {
   test: {
-    testMatch: [
-      '<rootDir>/packages/clippy/**/*.test.js',
-      '<rootDir>/packages/clippy/**/*.test.jsx',
-    ],
     name: 'clippy',
-    setupFiles: [
-      '../../config/setup/clippy.setup.js',
-    ],
-    environment: "happy-dom",
-  }
+    globals: true,
+    setupFiles: ['../../config/setup/clippy.setup.js'],
+    environment: 'happy-dom',
+    coverage: {
+      provider: 'istanbul',
+    },
+  },
 };