+++ packages/clippy/src/index.js
@@ -1,5 +1,0 @@
-import AGENTS from './agents';
-import useClippy from './useClippy';
-import ClippyProvider from './ClippyProvider';
-
-export { useClippy, AGENTS, ClippyProvider };