+++ packages/clippy/src/index.ts
@@ -1,0 +1,5 @@
+import AGENTS from './agents';
+import { ClippyProvider } from './ClippyProvider';
+import { useClippy } from './useClippy';
+
+export { useClippy, AGENTS, ClippyProvider };