+++ packages/clippy/src/useClippy.js
@@ -1,6 +1,0 @@
-import { useContext } from 'react';
-import ClippyContext from './ClippyContext';
-
-const useClippy = () => useContext(ClippyContext);
-
-export default useClippy;