+++ packages/clippy/src/useClippy.ts
@@ -1,0 +1,4 @@
+import { useContext } from 'react';
+import { ClippyContext } from './ClippyContext';
+
+export const useClippy = () => useContext(ClippyContext);